+++ GroceryDeliveryClient/src/pages/index/index-page.tsx
@@ -1,6 +1,6 @@
 import type { FC } from 'react';
 import { Sidebar, SidebarContent, SidebarGroup, SidebarGroupContent, SidebarHeader, SidebarInset, SidebarMenu, SidebarMenuButton, SidebarMenuItem, SidebarProvider, SidebarRail } from '@/components/ui/sidebar';
-import { useIsMobile } from '@/hooks/use-mobile';
+import { useIsMobile } from '@/hooks/mobile';
 import { ProductCard } from '@/components/product-card';
 import type { Product } from '@/types';
